% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{auxiliary_functions}
\alias{auxiliary_functions}
\title{Auxiliary Functions of a Graph Partition}
\usage{
auxiliary_functions(g, com, edgelist)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object)}

\item{com}{Community membership vector. Each element corresponds to a vertex
of the graph, and contains the index of the community it belongs to.}

\item{edelist}{alternatively, the edgelist of the graph}
}
\description{
Given a weighted graph and a partition into communities, returns the 
internal edge weight, the size, and the cut weight for each community.
}
\keyword{internal}
