% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{average_degree}
\alias{average_degree}
\title{Average Degree}
\usage{
average_degree(g, com)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object). If the edges have a "weight"
attribute, those will be used as weights.}

\item{com}{community membership integer vector. Each element corresponds to a vertex.}
}
\value{
Numeric vector with the average degree of each community.
}
\description{
Average degree (weighted degree, if the graph is weighted) of a graph's communities.
}
\examples{
data(karate, package="igraphdata")
average_degree(karate, membership(cluster_louvain(karate)))
}
