% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{average_odf}
\alias{average_odf}
\title{Average Out Degree Fraction}
\usage{
average_odf(
  g,
  com,
  no_clustering_coef = TRUE,
  type = "local",
  weighted = TRUE,
  w_max = NULL
)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object). If the edges have a "weight" attribute, those will be used as weights (otherwise, all edges are assumed to be 1).}

\item{com}{Community membership integer vector. Each element corresponds to a vertex
of the graph, and contains the index of the community it belongs to.}
}
\value{
Numeric vector with the Average ODF of each community.
}
\description{
Computes the Average Out Degree Fraction (Average ODF) of a graph (which can be weighted)
and its communities.
}
\examples{
data(karate, package="igraphdata")
average_odf(karate, membership(cluster_louvain(karate)))
}
