% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barabasi_albert.R
\name{barabasi_ablert_blocks}
\alias{barabasi_ablert_blocks}
\title{Generates a Barabási-Albert graph with community structure}
\usage{
barabasi_ablert_blocks(
  m,
  p,
  B,
  G0 = NULL,
  G0_labels = NULL,
  t_max,
  sample_with_replacement = FALSE
)
}
\arguments{
\item{m}{number of edges added at each step.}

\item{p}{vector of label probabilities}

\item{B}{matrix indicating the affinity of vertices of each label.}

\item{G0}{initial graph}

\item{G0_labels}{labels of the initial graph. If NULL, they will al be set to 1.}

\item{t_max}{final graph order}

\item{sample_with_replacement}{If TRUE, allows parallel edges.}
}
\description{
Generates a Barabási-Albert graph with community structure
}
