% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barabasi_albert.R
\name{barabasi_albert_blocks}
\alias{barabasi_albert_blocks}
\title{Generates a Barabási-Albert graph with community structure}
\usage{
barabasi_albert_blocks(
  m,
  p,
  B,
  G0 = NULL,
  G0_labels = NULL,
  t_max,
  sample_with_replacement = FALSE,
  type = "Hajek"
)
}
\arguments{
\item{m}{number of edges added at each step.}

\item{p}{vector of label probabilities. If they don't sum 1, they will be scaled accordingly.}

\item{B}{matrix indicating the affinity of vertices of each label.}

\item{G0}{initial graph}

\item{G0_labels}{labels of the initial graph. If NULL, they will al be set to 1.}

\item{t_max}{maximum value of t (which corresponds to graph order)}

\item{sample_with_replacement}{If TRUE, allows parallel edges.}

\item{type}{Either "Hajek" or "block_first".}

\item{t0}{initial t}
}
\value{
The resulting graph, as an igraph object. The vertices have a
"label" attribute.
}
\description{
Generates a Barabási-Albert graph with community structure
}
