% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{boot_alg_list}
\alias{boot_alg_list}
\title{Performs nonparametric bootstrap to a graph and a list of clustering algorithms}
\usage{
boot_alg_list(alg_list, g, R = 999, return_data = FALSE, type = "global")
}
\arguments{
\item{alg_list}{List of igraph clustering algorithms}

\item{g}{igraph graph object}

\item{R}{Number of bootstrap replicates.}

\item{type}{Can be "global" (Variation of Information, Reduced Mutual Information, 
and adjusted Rand Index) or "cluster-wise" (Jaccard distance)}
}
\description{
Performs nonparametric bootstrap on a graph's by resampling its vertices and
clustering the results using a list of clustering algorithms.
}
