% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_rs_table}
\alias{c_rs_table}
\title{Contingency table from membership vectors}
\usage{
c_rs_table(c1, c2)
}
\arguments{
\item{c1, c2}{membership vectors (integer values containing the index of each community)}
}
\description{
Given two membership vectors, returns the corresponding contingency table.
we assume the labels are >=1 and numbered consecutively. If not consecutive
(some labels are unused) this implementation still works, but will be less
efficient.
}
\keyword{internal}
