% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustAnalytics-package.R
\docType{package}
\name{clustAnalytics-package}
\alias{clustAnalytics}
\alias{clustAnalytics-package}
\title{clustAnalytics: Cluster Evaluation on Graphs}
\description{
Evaluates the stability and significance of clusters on 'igraph' graphs. Supports weighted and unweighted graphs. Implements the cluster evaluation methods defined by Arratia A, Renedo M (2021) \doi{10.7717/peerj-cs.600}. Also includes an implementation of the Reduced Mutual Information introduced by Newman et al. (2020) \doi{10.1103/PhysRevE.101.042304}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/martirm/clustAnalytics}
  \item Report bugs at \url{https://github.com/martirm/clustAnalytics/issues}
}

}
\keyword{internal}
