% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustAnalytics.R
\docType{package}
\encoding{UTF-8}
\name{clustAnalytics}
\alias{clustAnalytics}
\title{clustAnalytics}
\description{
This package evaluates the stability and significance of clusters in 
\code{igraph} graphs. Supports weighted and unweighted graphs.
}
\author{
Martí Renedo Mirambell
}
