% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables_auxiliary_functions.R
\name{contingency_to_membership_vectors}
\alias{contingency_to_membership_vectors}
\title{Computes possible membership vectors from contingency table}
\usage{
contingency_to_membership_vectors(M)
}
\arguments{
\item{M}{the contingency table}
}
\value{
a list containing the two membership vectors
}
\description{
Given a contingency table, obtains a possible pair of corresponding labelings.
That is, element M[i,j] is the number of elements that belong to community i
in the first labeling and j in the second.
}
