% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{density_ratio}
\alias{density_ratio}
\title{Density Ratio}
\usage{
density_ratio(g, com, weighted = TRUE, w_max = NULL, type = "local")
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object). If the edges have a "weight"
attribute, those will be used as weights.}

\item{com}{community membership integer vector. Each element corresponds to a vertex.}

\item{type}{can either be "local" or "global"}
}
\value{
Numeric vector with the internal density of each community.
}
\description{
Internal density of a graph's communities.
}
