% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{edges_inside}
\alias{edges_inside}
\title{Edges Inside}
\usage{
edges_inside(g, com)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object). If the edges have a "weight"
attribute, those will be used as weights.}

\item{com}{community membership integer vector. Each element corresponds to a vertex.}
}
\value{
Numeric vector with the internal edge weight of each community
}
\description{
Number of edges inside a graph's communities, or their accumulated weight if
the graph's edges are weighted.
}
\examples{
data(karate, package="igraphdata")
edges_inside(karate, membership(cluster_louvain(karate)))
}
\seealso{
Other cluster scoring functions: 
\code{\link{FOMD}()},
\code{\link{average_degree}()},
\code{\link{average_odf}()},
\code{\link{conductance}()},
\code{\link{coverage}()},
\code{\link{cut_ratio}()},
\code{\link{density_ratio}()},
\code{\link{expansion}()},
\code{\link{internal_density}()},
\code{\link{max_odf}()},
\code{\link{normalized_cut}()}
}
\concept{cluster scoring functions}
