% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_H_fraction}
\alias{estimate_H_fraction}
\title{Estimates |H_i|/|H_{i+1}|}
\usage{
estimate_H_fraction(M, k, l, error = 0.1)
}
\arguments{
\item{M}{matrix}

\item{k, l}{Coordinates of the first element that is not invariant}

\item{error}{error for the convergence of the method}
}
\value{
value of H_i/H_{i+1}
}
\description{
Estimates the fraction of elements of H_i that are also in H_{i+1} (where i=(k,l))
}
