% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_H_fraction_r_rows}
\alias{estimate_H_fraction_r_rows}
\title{Estimates |H_0|/|H_r*|}
\usage{
estimate_H_fraction_r_rows(M, r, error = 0.1)
}
\arguments{
\item{M}{contingency table}

\item{r}{row index}

\item{error}{error for the convergence of the method}
}
\description{
This is the total number of contingency tables (of the same margins as M) divided 
by the number that match M until the r-th row (included, 0-indexed). Note that
if r==0, this is always 1 by definition.
}
