% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_H_fractions}
\alias{estimate_H_fractions}
\title{Estimates |H_i|/|H_{i+1}| for the first r rows}
\usage{
estimate_H_fractions(M, r, error = 0.1)
}
\arguments{
\item{M}{contingency table}

\item{r}{row index}

\item{error}{error for the convergence of the method}
}
\value{
NumericVector containing all the ratios
}
\description{
The product of all these ratios is is the total number of contingency tables (of the same margins as M) divided 
by the number that match M until the r-th row (included, 0-indexed).
}
\keyword{internal}
