% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_functions.R
\name{evaluate_significance_r}
\alias{evaluate_significance_r}
\title{Computes scoring functions of a graph clustering}
\usage{
evaluate_significance_r(
  g,
  alg_list = clust_alg_list,
  no_clustering_coef = FALSE,
  ground_truth = FALSE,
  gt_clustering = NULL,
  table_style = 4,
  ignore_degenerate_cl = TRUE,
  Q = 100,
  lower_bound = 0,
  upper_bound = NULL,
  weight_sel = "const_var",
  n_reps = 1,
  w_max = 1,
  parallel = FALSE
)
}
\description{
Computes the scoring functions for both the graph and multiple randomized samples 
generated using a switching model
}
