% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{FOMD}
\alias{FOMD}
\title{FOMD (Fraction Over Median Degree)}
\usage{
FOMD(g, com, edgelist)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object)}

\item{com}{Community membership vector. Each element corresponds to a vertex}

\item{edelist}{alternatively, the edgelist of the graph}
}
\description{
Given a weighted graph and a partition into communities, returns the fraction
of nodes of each community whose internal degree (i.e. the degree accounting 
only intra-community edges) is greater than the median degree of the whole
graph.
}
