% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{igraph_to_edgelist}
\alias{igraph_to_edgelist}
\title{Returns edgelist with weights from a weighted igraph graph}
\usage{
igraph_to_edgelist(g, sort = TRUE)
}
\arguments{
\item{g}{igraph graph with weighted edges}

\item{sort}{sorts the edge list lexicographically before returning}
}
\value{
A matrix where the first two columns indicate the incident vertices,
and the third is the weight of the corresponding edge.
}
\description{
This function is just used internally for testing the package
}
