% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_mutual_information.R
\name{log_omega_estimation}
\alias{log_omega_estimation}
\title{Approximation of log(omega(a,b))}
\usage{
log_omega_estimation(c1, c2, base = exp(1))
}
\arguments{
\item{c1, c2}{membership vectors}

\item{base}{base of the logarithm (e by default)}
}
\description{
Where omega(a,b) is the number of contingency tables with a, b as row and column sums.
This approximation is only good for dense tables.
}
