% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_graph_weighted}
\alias{make_graph_weighted}
\title{Make graph weighted}
\usage{
make_graph_weighted(g)
}
\arguments{
\item{g}{igraph graph}
}
\value{
igraph graph with either all edge weights set to 1 (if the original
graph was unweighted), or to their original weights if they already existed
(in this case, the graph isn't modified at all).
}
\description{
Given a graph, create a "weight" attribute set to 1 for the edges if it doesn't
exist already.
}
