% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{max_odf}
\alias{max_odf}
\title{Max Out Degree Fraction}
\usage{
max_odf(g, com)

max_odf(g, com)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object). If the edges have a "weight" attribute, those will be used as weights (otherwise, all edges are assumed to be 1).}

\item{com}{Community membership integer vector. Each element corresponds to a vertex
of the graph, and contains the index of the community it belongs to.}
}
\value{
Numeric vector with the Max ODF of each community.

Numeric vector with the Max ODF of each community.
}
\description{
Computes the Maximum Out Degree Fraction (Max ODF) of a graph (which can be weighted)
and its communities.

Computes the Flake Out Degree Fraction (Max ODF) of a graph (which can be weighted)
and its communities.
}
\examples{
data(karate, package="igraphdata")
max_odf(karate, membership(cluster_louvain(karate)))
data(karate, package="igraphdata")
max_odf(karate, membership(cluster_louvain(karate)))
}
