% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_contingency_tables.R
\name{MC_test}
\alias{MC_test}
\title{Test for contingency table approximate counting}
\usage{
MC_test(M)
}
\description{
Count contingency tables with a single MC sampling process
}
\keyword{internal}
