% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{relabel}
\alias{relabel}
\title{Relabels membership vector}
\usage{
relabel(c)
}
\arguments{
\item{c}{numeric vector of vertex ids, not necessarily consecutive}
}
\value{
A numeric vector of consecutive vertex ids starting from one
}
\description{
Takes a vector of vertex ids indicating community membership, and relabels the communities
to have consecutive values from 1 to the number of communities.
}
