% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables_auxiliary_functions.R
\name{sort_matrix}
\alias{sort_matrix}
\title{Sort matrix}
\usage{
sort_matrix(M)
}
\arguments{
\item{M}{matrix}
}
\value{
rearranged matrix
}
\description{
Given a matrix, rearranges rows and columns so that row sums and col sums
end up in ascending order.
}
