% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{triangle_participation_ratio_communities}
\alias{triangle_participation_ratio_communities}
\title{Triangle Participation Ratio (community-wise)}
\usage{
triangle_participation_ratio_communities(g, com)
}
\arguments{
\item{g}{The input graph (as an igraph object). Edge weights and directions are ignored.}

\item{com}{Community membership vector. Each element corresponds to a vertex
of the graph, and contains the index of the community it belongs to.}
}
\value{
A vector containing the triangle participation ratio of each community.
}
\description{
Computes the triangle participation ratio (proportion of vertices that belong 
to a triangle). The computation is done to the subgraphs induced by each of 
the communities in the given partition.
}
