% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{walk_step}
\alias{walk_step}
\title{Performs a step of the Markov Chain Monte Carlo method}
\usage{
walk_step(M, k, l)
}
\arguments{
\item{M}{matrix}

\item{k, l}{Coordinates of the first element that is not invariant}
}
\value{
boolean indicating whether the step left the matrix invariant
}
\description{
Modifies the matrix while keeping the column and row sums constant, as well as
leaving the positions strictly preceding (k,l) in lexicographical order invariant.
}
\keyword{internal}
