% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{weighted_transitivity}
\alias{weighted_transitivity}
\title{Weighed transitivity of a weighted graph.}
\usage{
weighted_transitivity(g, upper_bound = NULL)
}
\arguments{
\item{g}{igraph graph}

\item{upper_bound}{upper bound to the edge weights used to compute the integral}
}
\value{
The weighted transitivity of the graph (a scalar).
}
\description{
Computed using the definition given by McAssey, M. P. and Bijma, F. in 
"A clustering coefficient for complete weighted networks" (2015).
}
