% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_functions.R
\name{evaluate_significance}
\alias{evaluate_significance}
\title{Evaluates significance of cluster algorithm results on a graph}
\usage{
evaluate_significance(
  g,
  alg_list = clust_alg_list,
  no_clustering_coef = TRUE,
  ground_truth = FALSE,
  gt_clustering = NULL,
  w_max = 1
)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object)}

\item{alg_list}{list of clustering algorithms, which take an \code{igraph} graph as
input and return an object of the \code{communities} class.}

\item{no_clustering_coef}{if \code{TRUE}, skips the computation of the clustering
coefficient, which is the most computationally costly of the scoring functions.}

\item{ground_truth}{if set to\code{TRUE}, computes the scoring functions for 
a ground truth clustering, which has to be provided as \code{gt_clustering}}

\item{gt_clustering}{ground truth clustering. Only used if \code{ground_truth}
is set to \code{TRUE}.}
}
\description{
Given a graph and a list of clustering algorithms, computes several scoring
functions on the clusters found by each of the algorithms.
}
