% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_mutual_information.R
\name{reduced_mutual_information}
\alias{reduced_mutual_information}
\title{Corrected Mutual Information}
\usage{
reduced_mutual_information(
  c1,
  c2,
  base = exp(1),
  normalized = FALSE,
  relabel = TRUE
)
}
\arguments{
\item{c1, c2}{membership vectors}

\item{base}{base of the logarithms used in the calculations. Changing it only scales the final value. By default set to e=exp(1).}

\item{normalized}{If true, computes the normalized version of the corrected mutual information.}
}
\description{
Computes the Newman's corrected mutual information (insert Latex reference here)
}
\details{
The implementation is based on equations 23 (25 for the normalized case) and 29 in (reference here).
The evaluations of the \eqn{\Gamma} functions can get too large and cause overflow 
issues in the intermediate steps, so the following term of equation 29:
\deqn{\frac{1}{2} \log \frac{\Gamma(\mu R) \Gamma(\nu S)} {(\Gamma(\nu)\Gamma(R))^S (\Gamma(\mu)\Gamma(S))^R } }
is rewritten as 
\deqn{\frac{1}{2} (\log\Gamma(\mu R) + \log\Gamma(\nu S) - S\log(\Gamma(\nu) - S\log(\Gamma(R) - R\log\Gamma(\mu) - R\log\Gamma(R)  )},
and then the function \link[base]{lgamma} is used instead of \link[base]{gamma}.
}
