% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions_Rcpp.R
\name{scoring_functions}
\alias{scoring_functions}
\title{Scoring Functions of a Graph Partition}
\usage{
scoring_functions(
  g,
  com,
  ignore_NA = TRUE,
  no_clustering_coef = TRUE,
  type = "local",
  weighted = TRUE
)
}
\arguments{
\item{g}{Graph to be analyzed (as an igraph object)}

\item{com}{Community membership vector. Each element corresponds to a vertex
of the graph, and contains the index of the community it belongs to.}

\item{type}{can be "local" for a cluster by cluster analysis, or "global" for
a global analysis of the whole graph partition.}
}
\value{
If \code{type=="local"}, returns a dataframe with a row for each 
community, and a column for each score. If \code{type=="global"}, returns a
single row with the weighted average scores.
}
\description{
Computes the scoring functions of a graph and its clusters
}
\examples{
scoring_functions_Rcpp(karate, membership(cluster_louvain(karate)))
}
