% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{weighted_clustering_coefficient}
\alias{weighted_clustering_coefficient}
\title{Clustering coefficient of a graph.}
\usage{
weighted_clustering_coefficient(g, upper_bound = NULL)
}
\arguments{
\item{g}{igraph graph}

\item{upper_bound}{upper bound to the edge weights used to compute the integral}
}
\description{
Computed using the definition given by McAssey, M. P. and Bijma, F. in 
"A clustering coefficient for complete weighted networks" (2015).
}
